# doc/openapi.yaml
...
openapi: 3.0.3
info:
  title: OTUS homework API
  contact:
    name: OTUS homework API Support
    url: https://www.facebook.com/vladimir.radchuk
    email: radchuk@hotmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
security:
- default: []
...
paths:
  /accounts/self:
    get:
      tags:
      - Accounts API
      summary: Returns the amount of money on the account
      parameters:
      - name: userId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: 'The request has been successfully applied '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmountDto'
        "400":
          description: The request has not been applied because of invalid input parameters.
        "401":
          description: The request has not been applied because it lacks valid authentication
            credentials for the target resource.
        "403":
          description: "The request is understood, but it has been refused or access\
            \ is not allowed. 1) The code is returned in case of an issue exception.\
            \ Please see RestIssueLog in the body 2) The user is suspended. The message\
            \ in the body exists."
        "404":
          description: "The URI requested is invalid or the resource requested, does\
            \ not exist."
        "500":
          description: Internal Server Error. The service call has not succeeded.
            The string in the body may contain the details.
    put:
      tags:
      - Accounts API
      summary: Updates the amount of money (sum with what exists)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmountDto'
      responses:
        "200":
          description: 'The request has been successfully applied '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmountDto'
        "400":
          description: The request has not been applied because of invalid input parameters.
        "401":
          description: The request has not been applied because it lacks valid authentication
            credentials for the target resource.
        "403":
          description: "The request is understood, but it has been refused or access\
            \ is not allowed. 1) The code is returned in case of an issue exception.\
            \ Please see RestIssueLog in the body 2) The user is suspended. The message\
            \ in the body exists."
        "404":
          description: "The URI requested is invalid or the resource requested, does\
            \ not exist."
        "500":
          description: Internal Server Error. The service call has not succeeded.
            The string in the body may contain the details.
...
  /events/orders/{id}:
    get:
      tags:
      - Events API
      summary: Returns all events for the specific order
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 'The request has been successfully applied '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderEventDto'
        "400":
          description: The request has not been applied because of invalid input parameters.
        "401":
          description: The request has not been applied because it lacks valid authentication
            credentials for the target resource.
        "403":
          description: "The request is understood, but it has been refused or access\
            \ is not allowed. 1) The code is returned in case of an issue exception.\
            \ Please see RestIssueLog in the body 2) The user is suspended. The message\
            \ in the body exists."
        "404":
          description: "The URI requested is invalid or the resource requested, does\
            \ not exist."
        "500":
          description: Internal Server Error. The service call has not succeeded.
            The string in the body may contain the details.
    post:
      tags:
      - Events API
      summary: Submits the order for billing
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DescriptionDto'
      responses:
        "200":
          description: 'The request has been successfully applied '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefDto'
        "401":
          description: The request has not been applied because it lacks valid authentication
            credentials for the target resource.
        "403":
          description: "The request is understood, but it has been refused or access\
            \ is not allowed. 1) The code is returned in case of an issue exception.\
            \ Please see RestIssueLog in the body 2) The user is suspended. The message\
            \ in the body exists."
        "404":
          description: "The URI requested is invalid or the resource requested, does\
            \ not exist."
        "500":
          description: Internal Server Error. The service call has not succeeded.
            The string in the body may contain the details.
    delete:
      tags:
      - Events API
      summary: Cancels the order
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 'The request has been successfully applied '
        "400":
          description: The request has not been applied because of invalid input parameters.
        "401":
          description: The request has not been applied because it lacks valid authentication
            credentials for the target resource.
        "403":
          description: "The request is understood, but it has been refused or access\
            \ is not allowed. 1) The code is returned in case of an issue exception.\
            \ Please see RestIssueLog in the body 2) The user is suspended. The message\
            \ in the body exists."
        "404":
          description: "The URI requested is invalid or the resource requested, does\
            \ not exist."
        "500":
          description: Internal Server Error. The service call has not succeeded.
            The string in the body may contain the details.
...
  /orders:
    get:
      tags:
      - Orders API
      summary: Returns all orders for a specific user
      parameters:
      - name: sort
        in: query
        schema:
          type: string
      - name: state
        in: query
        schema:
          type: string
      responses:
        "200":
          description: 'The request has been successfully applied '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
        "400":
          description: The request has not been applied because of invalid input parameters.
        "401":
          description: The request has not been applied because it lacks valid authentication
            credentials for the target resource.
        "403":
          description: "The request is understood, but it has been refused or access\
            \ is not allowed. 1) The code is returned in case of an issue exception.\
            \ Please see RestIssueLog in the body 2) The user is suspended. The message\
            \ in the body exists."
        "404":
          description: "The URI requested is invalid or the resource requested, does\
            \ not exist."
        "500":
          description: Internal Server Error. The service call has not succeeded.
            The string in the body may contain the details.
  /orders/{id}:
    get:
      tags:
      - Orders API
      summary: Returns all orders for a specific user
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 'The request has been successfully applied '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
        "400":
          description: The request has not been applied because of invalid input parameters.
        "401":
          description: The request has not been applied because it lacks valid authentication
            credentials for the target resource.
        "403":
          description: "The request is understood, but it has been refused or access\
            \ is not allowed. 1) The code is returned in case of an issue exception.\
            \ Please see RestIssueLog in the body 2) The user is suspended. The message\
            \ in the body exists."
        "404":
          description: "The URI requested is invalid or the resource requested, does\
            \ not exist."
        "500":
          description: Internal Server Error. The service call has not succeeded.
            The string in the body may contain the details.
...
components:
  schemas:
    AmountDto:
      description: DTO represents the amount of money
      type: object
      properties:
        amount:
          description: Amount of money
          type: number
...
    OrderDto:
      description: DTO represents normalized order
      type: object
      properties:
        creationDate:
          format: date-time
          description: order's creation date
          type: string
        description:
          description: order's description
          type: string
        id:
          description: order's identifier
          type: string
        items:
          allOf:
          - $ref: '#/components/schemas/ListItemDto'
          - description: order's items
        state:
          description: order's current state
          type: string
        userId:
          description: user who created an order
          type: string
    ItemDto:
      description: DTO represents single order item
      type: object
      properties:
        itemId:
          description: item's identifier
          type: string
        itemName:
          description: item's name
          type: string
        price:
          description: item's price
          type: number
        quantity:
          format: int32
          description: item's quantity
          type: integer
    ListItemDto:
      type: array
      items:
        $ref: '#/components/schemas/ItemDto'
    RefDto:
      description: DTO with reference information about dependent object
      type: object
      properties:
        id:
          description: identifier for referenced object
          type: string
    OrderEventDto:
      description: DTO represents information about order event
      type: object
      properties:
        creationDate:
          format: date-time
          description: event's timestamp
          type: string
        data:
          description: event's data
          type: string
        externalId:
          description: order's identifier
          type: string
        type:
          description: type of the event
          type: string
        userId:
          description: user who created the event
          type: string
    DescriptionDto:
      description: DTO represents information about order description
      type: object
      properties:
        description:
          description: description for the order
          type: string
...
  securitySchemes:
    default:
      type: apiKey
      description: Token that is returned after successfull login should be included
        in the cookie token for all requests to the API calls
      name: token
      in: cookie